+++ copy_this/modules/klarna/Requests/CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+Not Yet Released
+-----
+* Klarna Logs BugFix
+
 1.7.0
 -----